chore(viteroll): react example (#64)
