+++ viteroll/README.md
@@ -4,5 +4,5 @@ https://github.com/users/hi-ogawa/projects/4/views/1?pane=issue&itemId=84997064
 
 ```sh
 pnpm dev
-VITEROLL_HMR=false pnpm dev
+pnpm -C examples/react dev
 ```
