+++ viteroll/examples/react/tsconfig.json
@@ -5,10 +5,12 @@
 		"noUnusedLocals": true,
 		"noUnusedParameters": true,
 		"verbatimModuleSyntax": true,
+		"allowImportingTsExtensions": true,
 		"moduleResolution": "Bundler",
 		"module": "ESNext",
 		"target": "ESNext",
 		"lib": ["ESNext", "DOM"],
+		"jsx": "react-jsx",
 		"noEmit": true
 	}
 }
