+++ viteroll/examples/react/vite.config.ts
@@ -1,0 +1,10 @@
+import { defineConfig } from "vite";
+import { viteroll } from "../../viteroll";
+
+export default defineConfig({
+	plugins: [
+		viteroll({
+			reactRefresh: true,
+		}),
+	],
+});
