+++ viteroll/package.json
@@ -1,22 +1,33 @@
 {
-	"name": "@hiogawa/experiment-bundled-vite",
-	"private": true,
+	"name": "@hiogawa/viteroll",
+	"version": "0.0.0-pre.0",
 	"type": "module",
 	"scripts": {
 		"dev": "vite dev",
 		"lint": "biome check --write ."
 	},
 	"dependencies": {
-		"preact": "^10.24.3"
+		"magic-string": "^0.30.12",
+		"react-refresh": "^0.14.2",
+		"sirv": "^3.0.0"
 	},
 	"devDependencies": {
 		"@biomejs/biome": "^1.9.4",
 		"@types/node": "^22.8.5",
-		"magic-string": "^0.30.12",
+		"preact": "^10.24.3",
 		"rolldown": "file:rolldown-0.13.2.tgz",
-		"sirv": "^3.0.0",
 		"typescript": "^5.6.3",
 		"vite": "^5.4.10"
 	},
-	"packageManager": "pnpm@9.12.3+sha512.cce0f9de9c5a7c95bef944169cc5dfe8741abfb145078c0d508b868056848a87c81e626246cb60967cbd7fd29a6c062ef73ff840d96b3c86c40ac92cf4a813ee"
+	"peerDependencies": {
+		"rolldown": "*",
+		"vite": "*"
+	},
+	"packageManager": "pnpm@9.12.3+sha512.cce0f9de9c5a7c95bef944169cc5dfe8741abfb145078c0d508b868056848a87c81e626246cb60967cbd7fd29a6c062ef73ff840d96b3c86c40ac92cf4a813ee",
+	"pnpm": {
+		"overrides": {
+			"rolldown": "$rolldown",
+			"vite": "$vite"
+		}
+	}
 }
